Move new text conversion-specific classes to new package
