+++ deferred-resources/src/androidTest/java/com/backbase/deferredresources/text/FormattedDeferredTextJavaTest.java
@@ -1,0 +1,22 @@
+package com.backbase.deferredresources.text;
+
+import com.backbase.deferredresources.DeferredFormattedString;
+import com.backbase.deferredresources.DeferredText;
+import org.junit.Test;
+
+import static com.backbase.deferredresources.TestContext.getContext;
+import static com.google.common.truth.Truth.assertThat;
+
+public final class FormattedDeferredTextJavaTest {
+
+    // This test is impossible in Kotlin because Kotlin enforces spreading the original array when passed as a vararg
+    @Test
+    public void formatted_initializedWithArray_doesNotReflectLaterArrayChanges() {
+        Object[] formatArgs = new String[]{"Original"};
+        DeferredText deferred = new FormattedDeferredText(new DeferredFormattedString.Constant("%s text"), formatArgs);
+        formatArgs[0] = "Changed";
+
+        CharSequence resolved = deferred.resolve(getContext());
+        assertThat(resolved).isEqualTo("Original text");
+    }
+}
